refactor(openai): split monolithic client.go into focused files
client.go (1253→561 lines, -55%): decompose into single-responsibility files
within the same package openai:

- chat.go (317 lines): Chat Completions API — SendChat, request construction,
  HTTP plumbing, standardSink
- responses.go (217 lines): Responses API — fromResponsesAPIResponse,
  processOutputItem, block handlers, responsesSink
- metrics.go (185 lines): token accounting — accumulateUsage,
  calculateFinalMetrics, normalizeTokenCounts, logTokenThroughput
- tools.go (52 lines): tool schema conversion — toOpenAITools, toOpenAISchema
- client.go (561 lines): struct, constructor, options, shared utilities

All methods remain on *client receiver. No behavioral changes.
go test ./... passes all 65 packages including integration/e2e.

Also: remove dead Chatter() method from agentinternal/accessor.go
(zero callers — both Chat/Shutdown call a.raw.AsChatter() directly).
