+++ internal/infrastructure/llm/openai/client.go
@@ -4,36 +4,17 @@
 package openai
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
-	"runtime"
 	"strings"
 	"time"
 
 	"github.com/gosharplite/tell-me-go/internal/domain/llm"
 	"github.com/gosharplite/tell-me-go/internal/domain/ports"
-	"github.com/gosharplite/tell-me-go/internal/domain/tools"
 	"github.com/gosharplite/tell-me-go/internal/infrastructure/auth"
-	llmerr "github.com/gosharplite/tell-me-go/internal/infrastructure/llm/llmerr"
 )
-
-const maxResponseBytes = 10 * 1024 * 1024 // 10 MB safety limit for LLM responses (prevents OOM from malformed/malicious payloads)
-
-// defaultMaxTokens is the safe-floor budget used when both WithMaxTokens
-// and WithThinkingBudget are unset. Matches Anthropic's defaultMaxTokens
-// to close the residual silent-truncation hole that existed when
-// neither knob was set: OpenAI's chat/completions API defaults
-// max_tokens to 4096 when the field is omitted, which routinely
-// truncated large tool calls (the same failure class fixed for
-// Anthropic in commit 5031162c and for Gemini in commit 0495c6a3).
-//
-// Pinned by TestOpenAI_DefaultMaxTokens_IsGenerous in maxtokens_test.go
-// and by TestOpenAI_WithMaxTokens_ZeroAndNoThinkingBudget_FallsBackToDefault.
-const defaultMaxTokens = 16384
 
 // client implements the llm.LLMClient interface for OpenAI-compatible APIs.
 type client struct {
@@ -322,325 +303,9 @@ type completionTokensDetails struct {
 	ReasoningTokens int32 `json:"reasoning_tokens"`
 }
 
-// apiStrategy bundles the decisions made when selecting between the
-// standard /chat/completions and the /responses API surfaces.
-type apiStrategy struct {
-	useResponses bool
-	effort       string
-	hasEffort    bool
-}
-
-// resolveAPIStrategy decides between the standard and /responses API
-// surfaces based on model capabilities and the presence of tools with
-// a reasoning-effort header.
-func (c *client) resolveAPIStrategy(toolCount int) apiStrategy {
-	effort, hasEffort := c.headers["reasoning_effort"]
-	useResponses := c.capabilities.RequiresResponsesAPI && toolCount > 0 && hasEffort
-	return apiStrategy{useResponses: useResponses, effort: effort, hasEffort: hasEffort}
-}
-
-// buildRequestBody assembles the chat request payload, populating
-// either Input (for /responses) or Messages (for /chat/completions)
-// depending on the selected API strategy.
-func (c *client) buildRequestBody(ctx context.Context, strat apiStrategy, history []*llm.Content, toolDecls []*tools.ToolDeclaration, resolver llm.AssetResolver) (*chatRequest, error) {
-	req := chatRequest{
-		Model: c.model,
-		Tools: c.toOpenAITools(toolDecls, strat.useResponses),
-	}
-
-	if strat.useResponses {
-		items, err := c.toResponsesInput(ctx, history, resolver)
-		if err != nil {
-			return nil, err
-		}
-		req.Input = items
-		req.Reasoning = &reasoningConfig{Effort: strat.effort}
-	} else {
-		messages, err := c.toStandardMessages(ctx, history, resolver)
-		if err != nil {
-			return nil, err
-		}
-		req.Messages = messages
-		if strat.hasEffort && c.capabilities.SupportsReasoningEffort {
-			req.ReasoningEffort = strat.effort
-		}
-	}
-
-	return &req, nil
-}
-
-// applyOutputBudget resolves and sets the appropriate max-tokens field
-// on the request, accounting for the /responses endpoint override.
-func (c *client) applyOutputBudget(req *chatRequest, useResponses bool) {
-	budget := c.resolveOutputBudget()
-	field := c.capabilities.MaxTokensField
-	if useResponses {
-		field = llm.MaxTokensFieldOutput
-	}
-	switch field {
-	case llm.MaxTokensFieldOutput:
-		req.MaxOutputTokens = budget
-	case llm.MaxTokensFieldCompletion:
-		req.MaxCompletionTokens = budget
-	case llm.MaxTokensFieldLegacy:
-		req.MaxTokens = budget
-	}
-}
-
-// injectTransportHints adds transport-specific kwargs required by
-// certain API surfaces (e.g., Vertex AI thinking mode).
-func (c *client) injectTransportHints(req *chatRequest) {
-	if c.capabilities.RequiresVertexThinkingKwargs {
-		req.ChatTemplateKwargs = map[string]any{"thinking": true}
-	}
-}
-
-// prepareChatRequest constructs the chat request payload by delegating
-// to focused helpers for API strategy, body assembly, output budget,
-// and transport hints.
-func (c *client) prepareChatRequest(ctx context.Context, history []*llm.Content, toolDecls []*tools.ToolDeclaration, resolver llm.AssetResolver) (*chatRequest, error) {
-	strat := c.resolveAPIStrategy(len(toolDecls))
-
-	req, err := c.buildRequestBody(ctx, strat, history, toolDecls, resolver)
-	if err != nil {
-		return nil, err
-	}
-
-	c.applyOutputBudget(req, strat.useResponses)
-	c.injectTransportHints(req)
-
-	return req, nil
-}
-
-// resolveOutputBudget implements the three-tier resolution rule
-// documented on WithMaxTokens:
-//  1. WithMaxTokens(N) where N > 0 → N.
-//  2. WithMaxTokens(0) or unset, WithThinkingBudget(M) where M > 0 → M.
-//  3. Both unset → defaultMaxTokens.
-//
-// The thinking-budget fallback at tier 2 is intentional and load-
-// bearing: it preserves byte-identical request payloads for
-// deployments that previously relied on THINKING_BUDGET to drive the
-// max-tokens field, accepted by the architect during Task H design
-// review (Decision 5/7 reconciliation).
-func (c *client) resolveOutputBudget() int {
-	budget := c.maxTokens
-	if budget == 0 {
-		budget = c.thinkingBudget
-	}
-	if budget == 0 {
-		budget = defaultMaxTokens
-	}
-	return budget
-}
-
-func (c *client) resolveEndpoint(req *chatRequest) string {
-	if c.capabilities.RequiresResponsesAPI && len(req.Tools) > 0 && (req.Reasoning != nil || req.ReasoningEffort != "") {
-		return "/responses"
-	}
-	return "/chat/completions"
-}
-
-func (c *client) createHTTPRequest(ctx context.Context, payload *chatRequest) (*http.Request, error) {
-	body, err := json.Marshal(payload)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal request: %w", err)
-	}
-
-	endpoint := c.resolveEndpoint(payload)
-	req, err := http.NewRequestWithContext(ctx, "POST", c.baseURL+endpoint, bytes.NewBuffer(body))
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	// Apply custom headers
-	for k, v := range c.headers {
-		req.Header.Set(k, v)
-	}
-
-	// Apply authentication
-	authReq := &auth.Request{Headers: make(map[string]string)}
-	if err := c.authenticator.Apply(ctx, authReq); err != nil {
-		return nil, err
-	}
-	for k, v := range authReq.Headers {
-		req.Header.Set(k, v)
-	}
-
-	return req, nil
-}
-
-func (c *client) SendChat(ctx context.Context, history []*llm.Content, toolDecls []*tools.ToolDeclaration, resolver llm.AssetResolver) (*llm.Content, *llm.Metrics, error) {
-	reqPayload, err := c.prepareChatRequest(ctx, history, toolDecls, resolver)
-	if err != nil {
-		return nil, nil, err
-	}
-	endpoint := c.resolveEndpoint(reqPayload)
-	req, err := c.createHTTPRequest(ctx, reqPayload)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	startTime := time.Now()
-	resp, err := c.httpClient.Do(req)
-	ttfb := time.Since(startTime) // Time To First Byte
-
-	if err != nil {
-		return nil, nil, fmt.Errorf("request failed: %w", err)
-	}
-	defer func() {
-		_ = resp.Body.Close()
-	}()
-
-	if resp.StatusCode != http.StatusOK {
-		bodyBytes, err := io.ReadAll(io.LimitReader(resp.Body, maxResponseBytes))
-		if err != nil {
-			return nil, nil, fmt.Errorf("api returned status %d; additionally, failed to read response body: %w", resp.StatusCode, err)
-		}
-		return nil, nil, &llmerr.APIError{
-			Status: resp.StatusCode,
-			Body:   string(bodyBytes),
-		}
-	}
-
-	// Stream the JSON decoding to avoid large memory allocations
-	bodyReadStart := time.Now()
-
-	if endpoint == "/responses" {
-		var chatResp responsesAPIResponse
-		if err := json.NewDecoder(io.LimitReader(resp.Body, maxResponseBytes)).Decode(&chatResp); err != nil {
-			return nil, nil, fmt.Errorf("failed to decode response: %w", err)
-		}
-
-		bodyReadTime := time.Since(bodyReadStart)
-		totalDuration := time.Since(startTime)
-
-		c.logger.Debug("http_timing_breakdown",
-			"platform", runtime.GOOS,
-			"provider", "openai",
-			"model", c.model,
-			"ttfb_ms", ttfb.Milliseconds(),
-			"body_read_ms", bodyReadTime.Milliseconds(),
-			"total_ms", totalDuration.Milliseconds(),
-			"endpoint", endpoint,
-		)
-
-		return c.fromResponsesAPIResponse(&chatResp, totalDuration.Seconds())
-	}
-
-	var chatResp chatResponse
-	if err := json.NewDecoder(io.LimitReader(resp.Body, maxResponseBytes)).Decode(&chatResp); err != nil {
-		return nil, nil, fmt.Errorf("failed to decode response: %w", err)
-	}
-
-	bodyReadTime := time.Since(bodyReadStart)
-	totalDuration := time.Since(startTime)
-
-	c.logger.Debug("http_timing_breakdown",
-		"platform", runtime.GOOS,
-		"provider", "openai",
-		"model", c.model,
-		"ttfb_ms", ttfb.Milliseconds(),
-		"body_read_ms", bodyReadTime.Milliseconds(),
-		"total_ms", totalDuration.Milliseconds(),
-		"endpoint", endpoint,
-	)
-
-	return c.fromOpenAIResponse(&chatResp, totalDuration.Seconds())
-}
-
 type openaiSink interface {
 	AddMessage(role, text string, reasoning *string, toolCalls []toolCall)
 	AddToolResponse(id, response string)
-}
-
-type responsesSink struct {
-	client *client
-	items  []historyItem
-}
-
-func (s *responsesSink) AddMessage(role, text string, reasoning *string, toolCalls []toolCall) {
-	r := role
-	s.items = append(s.items, historyItem{
-		Type:    "message",
-		Role:    &r,
-		Content: []requestContentBlock{{Type: s.client.resolveBlockType(role), Text: text}},
-	})
-	for _, tc := range toolCalls {
-		cid := tc.ID
-		name := tc.Function.Name
-		args := tc.Function.Arguments
-		s.items = append(s.items, historyItem{
-			Type:      "function_call",
-			CallID:    &cid,
-			Name:      &name,
-			Arguments: &args,
-		})
-	}
-}
-
-func (s *responsesSink) AddToolResponse(id, response string) {
-	cid := id
-	out := response
-	s.items = append(s.items, historyItem{
-		Type:   "function_call_output",
-		CallID: &cid,
-		Output: &out,
-	})
-}
-
-type standardSink struct {
-	messages []message
-}
-
-func (s *standardSink) AddMessage(role, text string, reasoning *string, toolCalls []toolCall) {
-	s.messages = append(s.messages, message{
-		Role:             role,
-		Content:          text,
-		ReasoningContent: reasoning,
-		ToolCalls:        toolCalls,
-	})
-}
-
-func (s *standardSink) AddToolResponse(id, response string) {
-	s.messages = append(s.messages, message{
-		Role:       "tool",
-		ToolCallID: id,
-		Content:    response,
-	})
-}
-
-func (c *client) resolveBlockType(role string) string {
-	if role == "assistant" {
-		return "output_text"
-	}
-	return "input_text"
-}
-
-func (c *client) toResponsesInput(ctx context.Context, history []*llm.Content, resolver llm.AssetResolver) ([]historyItem, error) {
-	sink := &responsesSink{client: c}
-	personaInjected := c.maybeInjectInitialPersona(sink)
-
-	for _, h := range history {
-		if err := c.appendMessagesFromHistoryItem(ctx, sink, h, resolver, &personaInjected); err != nil {
-			return nil, err
-		}
-	}
-	return sink.items, nil
-}
-
-func (c *client) toStandardMessages(ctx context.Context, history []*llm.Content, resolver llm.AssetResolver) ([]message, error) {
-	sink := &standardSink{}
-	personaInjected := c.maybeInjectInitialPersona(sink)
-
-	for _, h := range history {
-		if err := c.appendMessagesFromHistoryItem(ctx, sink, h, resolver, &personaInjected); err != nil {
-			return nil, err
-		}
-	}
-	return sink.messages, nil
 }
 
 func (c *client) maybeInjectInitialPersona(sink openaiSink) (personaInjected bool) {
@@ -774,369 +439,12 @@ func (c *client) injectPersona(sink openaiSink, personaInjected *bool, role stri
 	}
 }
 
-func (c *client) toOpenAITools(decls []*tools.ToolDeclaration, flattened bool) []tool {
-	if len(decls) == 0 {
-		return nil
-	}
-	var res []tool
-	for _, d := range decls {
-		t := tool{
-			Type: "function",
-		}
-		if flattened {
-			t.Name = d.Name
-			t.Description = d.Description
-			t.Parameters = toOpenAISchema(d.Parameters)
-		} else {
-			t.Function = &functionDeclaration{
-				Name:        d.Name,
-				Description: d.Description,
-				Parameters:  toOpenAISchema(d.Parameters),
-			}
-		}
-		res = append(res, t)
-	}
-	return res
-}
-
-func toOpenAISchema(s *tools.Schema) *schema {
-	if s == nil {
-		return nil
-	}
-	res := &schema{
-		Type:        strings.ToLower(s.Type),
-		Description: s.Description,
-		Required:    s.Required,
-		Enum:        s.Enum,
-		Items:       toOpenAISchema(s.Items),
-	}
-	if s.Properties != nil {
-		res.Properties = make(map[string]*schema)
-		for k, v := range s.Properties {
-			res.Properties[k] = toOpenAISchema(v)
-		}
-	}
-	return res
-}
-
-func (c *client) fromResponsesAPIResponse(resp *responsesAPIResponse, duration float64) (*llm.Content, *llm.Metrics, error) {
-	content := &llm.Content{
-		Role: "model",
-	}
-
-	mergedUsage := resp.Usage
-
-	for _, out := range resp.Output {
-		c.accumulateUsage(&mergedUsage, out.Usage)
-		if err := c.processOutputItem(content, &out); err != nil {
-			return nil, nil, err
-		}
-	}
-
-	content.Validate()
-
-	return content, c.calculateFinalMetrics(mergedUsage, duration), nil
-}
-
-func (c *client) accumulateUsage(merged *usage, itemUsage *usage) {
-	if itemUsage == nil {
-		return
-	}
-	if merged.PromptTokens == 0 {
-		merged.PromptTokens = itemUsage.PromptTokens
-	}
-	if merged.CompletionTokens == 0 {
-		merged.CompletionTokens = itemUsage.CompletionTokens
-	}
-	if merged.InputTokens == 0 {
-		merged.InputTokens = itemUsage.InputTokens
-	}
-	if merged.OutputTokens == 0 {
-		merged.OutputTokens = itemUsage.OutputTokens
-	}
-}
-
-func (c *client) processOutputItem(content *llm.Content, out *responseOutputItem) error {
-	if out.Message != nil {
-		for _, cb := range out.Message.Content {
-			c.appendPartsFromBlock(content, cb)
-		}
-		if err := c.parseResponseToolCalls(out.Message.ToolCalls, content); err != nil {
-			return err
-		}
-	} else {
-		// Process as direct content block based on type
-		cb := contentBlock{
-			Type:       out.Type,
-			Text:       out.Text,
-			InputText:  out.InputText,
-			OutputText: out.OutputText,
-			Thought:    out.Thought,
-			Reasoning:  out.Reasoning,
-			Refusal:    out.Refusal,
-		}
-		c.appendPartsFromBlock(content, cb)
-
-		// Fallback for items that put blocks in a top-level array
-		for _, childCb := range out.Content {
-			c.appendPartsFromBlock(content, childCb)
-		}
-
-		// Top-level tool calls in output item
-		if err := c.parseResponseToolCalls(out.ToolCalls, content); err != nil {
-			return err
-		}
-
-		// Detection logic for top-level tool call (type: "call")
-		targetName := out.Name
-		targetArgs := out.Arguments
-		if out.Function != nil {
-			targetName = out.Function.Name
-			targetArgs = out.Function.Arguments
-		}
-		if targetName != "" {
-			_ = c.appendToolCall(content, out.ID, targetName, targetArgs)
-		}
-	}
-	return nil
-}
-
-func (c *client) calculateFinalMetrics(u usage, duration float64) *llm.Metrics {
-	promptTokens, completionTokens, totalTokens := c.normalizeTokenCounts(u)
-	thinkingTokens := c.resolveThinkingTokens(u)
-
-	// OpenAI-compatible providers (DeepSeek reasoner, OpenAI gpt-5/o-series)
-	// report completion_tokens as the SUM of final-content tokens and
-	// reasoning tokens (CoT). The reasoning_tokens field inside
-	// completion_tokens_details is a SUBSET, not an addend.
-	//
-	// Verified against the live deepseek-reasoner API on 2025-12-04:
-	//   prompt_tokens=16, completion_tokens=190, total_tokens=206,
-	//   reasoning_tokens=147   →   total = prompt + completion (NOT + reasoning)
-	//
-	// Downstream pricing.go::Calculate treats ResponseTokens and
-	// ThinkingTokens as disjoint and sums them. Subtract here so that
-	// invariant holds: ResponseTokens + ThinkingTokens == completion_tokens.
-	// Without this fix, reasoning tokens are billed twice (up to ~2×
-	// overcharge on heavy-reasoning turns) and the UI's "O:" output total
-	// is structurally wrong.
-	//
-	// See: docs/adr/2026-04-reasoning-token-accounting.md (ADR-023)
-	contentTokens := completionTokens
-	if thinkingTokens > 0 && contentTokens >= thinkingTokens {
-		contentTokens -= thinkingTokens
-	}
-
-	metrics := &llm.Metrics{
-		Model:          c.model,
-		PromptTokens:   promptTokens,
-		ResponseTokens: contentTokens,
-		TotalTokens:    totalTokens,
-		Duration:       duration,
-		CachedTokens:   c.resolveCachedTokens(u),
-		ThinkingTokens: thinkingTokens,
-		TrafficType:    c.resolveTrafficType(u),
-	}
-
-	c.logTokenThroughput(metrics)
-
-	return metrics
-}
-
-// normalizeTokenCounts resolves prompt, completion, and total tokens from dual-field API responses.
-func (c *client) normalizeTokenCounts(u usage) (prompt, completion, total int32) {
-	prompt = u.PromptTokens
-	if prompt == 0 {
-		prompt = u.InputTokens
-	}
-	completion = u.CompletionTokens
-	if completion == 0 {
-		completion = u.OutputTokens
-	}
-	total = u.TotalTokens
-	if total == 0 {
-		total = prompt + completion
-	}
-	return
-}
-
-// resolveCachedTokens extracts cached token count from OpenAI or DeepSeek response fields.
-func (c *client) resolveCachedTokens(u usage) int32 {
-	if u.PromptCacheHitTokens > 0 {
-		return u.PromptCacheHitTokens
-	}
-	if u.PromptTokensDetails != nil {
-		return u.PromptTokensDetails.CachedTokens
-	}
-	return 0
-}
-
 // resolveThinkingTokens extracts reasoning/thinking token count from completion details.
 func (c *client) resolveThinkingTokens(u usage) int32 {
 	if u.CompletionTokensDetails != nil {
 		return u.CompletionTokensDetails.ReasoningTokens
 	}
 	return 0
-}
-
-// resolveTrafficType determines traffic type via server metadata or header reflection override.
-func (c *client) resolveTrafficType(u usage) string {
-	// 1. Priority 1: Source of Truth (Response)
-	trafficType := ""
-	if u.ExtraProperties != nil && u.ExtraProperties.Google != nil {
-		trafficType = u.ExtraProperties.Google.TrafficType
-	}
-
-	// 2. Priority 2: Reflected Intent (Force override if header is 'priority')
-	for k, v := range c.headers {
-		normalizedK := strings.ReplaceAll(strings.ToLower(k), "_", "-")
-		if normalizedK == "x-vertex-ai-llm-shared-request-type" && strings.TrimSpace(strings.ToLower(v)) == "priority" {
-			return "ON_DEMAND_PRIORITY"
-		}
-	}
-
-	return trafficType
-}
-
-// logTokenThroughput emits diagnostic token throughput metrics.
-func (c *client) logTokenThroughput(metrics *llm.Metrics) {
-	if metrics.ResponseTokens > 0 && metrics.Duration > 0.1 {
-		tokensPerSec := float64(metrics.ResponseTokens) / metrics.Duration
-		c.logger.Debug("token_throughput",
-			"platform", runtime.GOOS,
-			"provider", "openai",
-			"model", metrics.Model,
-			"response_tokens", metrics.ResponseTokens,
-			"duration_sec", metrics.Duration,
-			"tokens_per_sec", tokensPerSec,
-			"cached_tokens", metrics.CachedTokens,
-		)
-	}
-}
-
-func (c *client) handleTextBlock(content *llm.Content, cb contentBlock) {
-	text := c.extractBlockText(cb.Text)
-	if text == "" {
-		text = cb.OutputText
-	}
-	if text == "" {
-		text = cb.InputText
-	}
-	if text != "" {
-		content.Parts = append(content.Parts, &llm.Part{Text: text})
-	}
-}
-
-func (c *client) handleThoughtBlock(content *llm.Content, cb contentBlock) {
-	val := cb.Thought
-	if val == "" {
-		val = cb.Reasoning
-	}
-	// Some models use type: "thought" with a "text" field
-	if val == "" && cb.Type == "thought" {
-		val = c.extractBlockText(cb.Text)
-	}
-	if val != "" {
-		content.Parts = append(content.Parts, &llm.Part{Text: val, IsThought: true})
-	}
-}
-
-func (c *client) handleToolUseBlock(content *llm.Content, cb contentBlock) {
-	if cb.Name != "" && cb.ID != "" {
-		content.Parts = append(content.Parts, &llm.Part{
-			FunctionCall: &llm.FunctionCall{
-				ID:   cb.ID,
-				Name: cb.Name,
-				Args: cb.Input,
-			},
-		})
-	}
-}
-
-func (c *client) handleRefusalBlock(content *llm.Content, cb contentBlock) {
-	if cb.Refusal != "" {
-		content.Parts = append(content.Parts, &llm.Part{Text: cb.Refusal})
-	}
-}
-
-func (c *client) appendPartsFromBlock(content *llm.Content, cb contentBlock) {
-	switch cb.Type {
-	case "text", "output_text", "input_text":
-		c.handleTextBlock(content, cb)
-	case "thought", "reasoning":
-		c.handleThoughtBlock(content, cb)
-	case "tool_use":
-		c.handleToolUseBlock(content, cb)
-	case "refusal":
-		c.handleRefusalBlock(content, cb)
-	default:
-		c.logger.Debug("unhandled content block type", "type", cb.Type)
-	}
-}
-
-func (c *client) extractBlockText(txt interface{}) string {
-	switch v := txt.(type) {
-	case string:
-		return v
-	case map[string]interface{}:
-		if val, ok := v["value"].(string); ok {
-			return val
-		}
-	}
-	return ""
-}
-
-func (c *client) fromOpenAIResponse(resp *chatResponse, duration float64) (*llm.Content, *llm.Metrics, error) {
-	if len(resp.Choices) == 0 {
-		return nil, nil, fmt.Errorf("no choices returned from api")
-	}
-	choice := resp.Choices[0]
-	msg := choice.Message
-
-	content := &llm.Content{
-		Role: "model",
-	}
-
-	c.parseResponseContent(msg.Content, content)
-
-	// Reasoning content (DeepSeek extension)
-	if msg.ReasoningContent != nil && *msg.ReasoningContent != "" {
-		content.Parts = append(content.Parts, &llm.Part{Text: *msg.ReasoningContent, IsThought: true})
-	}
-
-	if err := c.parseResponseToolCalls(msg.ToolCalls, content); err != nil {
-		return nil, nil, err
-	}
-
-	content.Validate() // Final boundary sanitization
-
-	metrics := c.calculateFinalMetrics(resp.Usage, duration)
-
-	// Detect output-budget truncation. OpenAI returns
-	// finish_reason="length" when the response was cut off because the
-	// max_tokens (or max_completion_tokens) cap was reached. The
-	// downstream impact is the same as Anthropic's max_tokens stop:
-	// tool-call argument JSON may be cut mid-string, partial reasoning
-	// may mislead the agent. Surfacing as an error lets the caller
-	// decide whether to retry with a larger budget.
-	//
-	// In OpenAI's case, mid-string truncation typically also makes the
-	// tool-call args JSON unparseable, which appendToolCall already
-	// surfaces as a JSON unmarshal error. This finish_reason check
-	// covers the residual class where truncation lands at a JSON
-	// object boundary AND between keys, producing a partial-but-
-	// well-formed args object.
-	if choice.FinishReason == "length" {
-		return content, metrics, fmt.Errorf(
-			"response truncated at max_tokens (finish_reason=%q): "+
-				"output budget was exhausted before the model "+
-				"finished. Increase the max_tokens budget or shorten "+
-				"the prompt/response",
-			choice.FinishReason,
-		)
-	}
-
-	return content, metrics, nil
 }
 
 func (c *client) parseResponseContent(rawContent interface{}, content *llm.Content) {
